requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
langdetect==1.0.9
pytest==8.3.2
pre-commit==3.8.0
